remove Item{S,C}ao, drop items directly to parcels

+++ src/server.cpp
@@ -3974,9 +3974,18 @@ void Server::ProcessData(u8 *data, u32 datasize, u16 peer_id)
 				ServerActiveObject *obj = wielditem->createSAO(&m_env, 0, pos);
 
 				if (obj == NULL) {
-					infostream<<"WARNING: item resulted in NULL object, "
-							<<"not placing onto map"
-							<<std::endl;
+					InventoryItem *nitem;
+					if (g_settings->getBool("infinite_inventory") == false) {
+						// Delete the right amount of items from the slot
+						InventoryList *ilist = player->inventory.getList("main");
+						nitem = ilist->changeItem(item_i,NULL);
+						// Send inventory
+						UpdateCrafting(peer_id);
+						SendInventory(peer_id);
+					}else{
+						nitem = wielditem->clone();
+					}
+					m_env.dropToParcel(p_over,nitem);
 				}else{
 					actionstream<<player->getName()<<" places "<<item->getName()
 							<<" at "<<PP(p_over)<<std::endl;
@@ -4090,9 +4099,18 @@ void Server::ProcessData(u8 *data, u32 datasize, u16 peer_id)
 				}
 
 				if (obj == NULL) {
-					infostream<<"WARNING: item resulted in NULL object, "
-							<<"not placing onto map"
-							<<std::endl;
+					InventoryItem *nitem;
+					if (g_settings->getBool("infinite_inventory") == false) {
+						// Delete the right amount of items from the slot
+						InventoryList *ilist = player->inventory.getList("main");
+						nitem = ilist->changeItem(item_i,NULL);
+						// Send inventory
+						UpdateCrafting(peer_id);
+						SendInventory(peer_id);
+					}else{
+						nitem = wielditem->clone();
+					}
+					m_env.dropToParcel(p_over,nitem);
 				}else{
 					actionstream<<player->getName()<<" places "<<item->getName()
 							<<" at "<<PP(p_over)<<std::endl;
@@ -4378,19 +4396,21 @@ void Server::ProcessData(u8 *data, u32 datasize, u16 peer_id)
 					) {
 						InventoryList *list = player->inventory.getList("discard");
 						InventoryItem *item = list->getItem(0);
-						if (g_settings->getBool("droppable_inventory")) {
-							v3f pos = player->getPosition();
-							pos.Y += BS;
-							v3f dir = v3f(0,0,BS);
-							dir.rotateXZBy(player->getYaw());
-							pos += dir;
-							ServerActiveObject *obj = item->createSAO(&m_env,0,pos);
-							if (obj)
-								m_env.addActiveObject(obj);
-						}
-						if (g_settings->getBool("infinite_inventory") == false) {
-							list->deleteItem(0);
-							SendInventory(player->peer_id);
+						if (item) {
+							if (g_settings->getBool("droppable_inventory")) {
+								v3f pos = player->getPosition();
+								pos.Y += BS;
+								v3f dir = v3f(0,0,BS);
+								dir.rotateXZBy(player->getYaw());
+								pos += dir;
+								v3s16 pp = floatToInt(pos,BS);
+								item = item->clone();
+								m_env.dropToParcel(pp,item);
+							}
+							if (g_settings->getBool("infinite_inventory") == false) {
+								list->deleteItem(0);
+								SendInventory(player->peer_id);
+							}
 						}
 					}
 				}
@@ -5530,88 +5550,17 @@ void Server::HandlePlayerHP(Player *player, s16 damage, s16 suffocate, s16 hunge
 		player->wet = 0;
 		player->blood = 0;
 
-		//TODO: Throw items around
+		// Throw items around
 		if (g_settings->getBool("death_drops_inv")) {
 			v3s16 bottompos = floatToInt(player->getPosition() + v3f(0,-BS/4,0), BS);
 			v3s16 p = bottompos + v3s16(0,1,0);
-			MapNode in_n = m_env.getMap().getNodeNoEx(p);
-			v3s16 droppos = p;
-			bool can_drop = false;
-			// if they're standing in an air-like node, drop inventory to a parcel
-			// otherwise they just lose it
-			v3s16 pp;
-			if (in_n.getContent() == CONTENT_PARCEL) {
-				can_drop = true;
-			}else if (m_env.searchNear(p,v3s16(3,3,3),CONTENT_PARCEL,&pp)) {
-				droppos = pp;
-				can_drop = true;
-			}else if (in_n.getContent() != CONTENT_LAVASOURCE && content_features(in_n.getContent()).buildable_to) {
-				while (!can_drop && p.Y > -30000) {
-					MapNode nn = m_env.getMap().getNodeNoEx(p+v3s16(0,-1,0));
-					if (nn.getContent() == CONTENT_PARCEL) {
-						droppos = p+v3s16(0,-1,0);
-						can_drop = true;
-					}else if (nn.getContent() == CONTENT_LAVASOURCE) {
-						break;
-					}else if (!content_features(nn.getContent()).buildable_to) {
-						droppos = p;
-						can_drop = true;
-						break;
-					}
-					p.Y--;
-				}
-			}
-
-			if (can_drop) {
-				MapNode nn = m_env.getMap().getNodeNoEx(droppos);
-				if (nn.getContent() != CONTENT_PARCEL) {
-					v3s16 pp;
-					if (m_env.searchNear(droppos,v3s16(3,3,3),CONTENT_PARCEL,&pp)) {
-						droppos = pp;
-						nn = m_env.getMap().getNodeNoEx(droppos);
-					}else{
-						nn.setContent(CONTENT_PARCEL);
-						core::list<u16> far_players;
-						sendAddNode(droppos, nn, 0, &far_players, 30);
-
-						/*
-							Add node.
-
-							This takes some time so it is done after the quick stuff
-						*/
-						core::map<v3s16, MapBlock*> modified_blocks;
-						{
-							MapEditEventIgnorer ign(&m_ignore_map_edit_events);
-
-							std::string p_name(player->getName());
-							m_env.getMap().addNodeAndUpdate(droppos, nn, modified_blocks, p_name);
-						}
-						/*
-							Set blocks not sent to far players
-						*/
-						for (core::list<u16>::Iterator i = far_players.begin(); i != far_players.end(); i++) {
-							u16 peer_id = *i;
-							RemoteClient *client = getClient(peer_id);
-							if (client==NULL)
-								continue;
-							client->SetBlocksNotSent(modified_blocks);
-						}
-					}
-				}
-
-				NodeMetadata *meta = m_env.getMap().getNodeMetadata(droppos);
-				if (meta) {
-					Inventory *inv = meta->getInventory();
-					if (inv) {
-						InventoryList *pl = inv->getList("0");
-						InventoryList *il = player->inventory.getList("main");
-						if (pl && il) {
-							for (u32 i=0; i<32; i++) {
-								InventoryItem *item = il->takeItem(i,99);
-								if (item)
-									pl->addItem(item);
-							}
-						}
+			{
+				InventoryList *il = player->inventory.getList("main");
+				if (il) {
+					for (u32 i=0; i<32; i++) {
+						InventoryItem *item = il->changeItem(i,NULL);
+						if (item)
+							m_env.dropToParcel(p,item);
 					}
 				}
 			}
